Back at it! I think I know how to write this how I want it now!
