+++ src/data/__init__.py
@@ -1,0 +1,3 @@
+from data import handler, object
+
+
