+++ src/main.py
@@ -1,15 +1,5 @@
-import object
+import data
 
-account = object.account.AccountObject()
-
-account.hash.remember_password("password")
-
-entry: object.entry.PrivateEntryObject = account.vault.create_private_entry(
-    "password",
-    "name",
-    "account",
-    {"hello": "world!"},
-    "note!"
+data.object.AccountObject.get_public_entry_by(
+    test="hello"
 )
-
-
