--- src/object/__init__.py
@@ -1,4 +1,0 @@
-from object import account, hash, vault, entry
-
-
-
