--- src/object/account.py
@@ -1,8 +1,0 @@
-from object import vault, hash
-
-
-class AccountObject:
-    def __init__(self):
-        self.vault = vault.VaultObject()
-        self.hash = hash.ScryptObject()
-
